chore: break out body of trade input into a separate component
